requirements.txt
Flask
Flask-Cors
requests
gunicorn
beautifulsoup4
lxml
python-whois

